use separate link when app already installer and only token is needed
