+++ src/components/error-boundary.tsx
@@ -1,13 +1,17 @@
 import React from 'react'
 import { Link, useRouteError } from 'react-router-dom'
-import { GitHubApiError } from '../services/github-error'
+import { GitHubApiError } from '../errors/github-error'
 import { useTranslation } from 'react-i18next'
+import { AuthenticationError } from '../errors/authentication-error'
 
 const ErrorBoundary: React.FunctionComponent = () => {
   const error = useRouteError()
   const { t } = useTranslation('error')
 
-  if (error instanceof GitHubApiError && error.status === 401) {
+  if (
+    error instanceof AuthenticationError ||
+    (error instanceof GitHubApiError && error.status === 401)
+  ) {
     return (
       <div>
         <h1>{t('Authentication error!')}</h1>
