+++ src/errors/authentication-error.ts
@@ -1,0 +1,8 @@
+export class AuthenticationError extends Error {
+  constructor(
+    message: string,
+    public readonly innerError?: Error
+  ) {
+    super(message)
+  }
+}
