+++ src/errors/github-app-auth-flow-error.ts
@@ -1,0 +1,15 @@
+export class GitHubAppAuthFlowError extends Error {
+  constructor(
+    message: string,
+    public readonly code: string,
+    public readonly description: string
+  ) {
+    super(message)
+
+    this.name = this.constructor.name
+
+    if (Error.captureStackTrace) {
+      Error.captureStackTrace(this, this.constructor)
+    }
+  }
+}
